requirements.txt
albumentations
easydict
einops
numpy
opencv_python
pathlib
pycocotools
PyYAML
scipy
setuptools
timm
torchvision
tqdm
loguru
py-cpuinfo
trimesh
open3d
onnx
onnxslim
onnxruntime-gpu
wheel
build
pandas


